requirements.txt
aiogram==2.11.2
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.10.0
astroid==2.14.1
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
Deprecated==1.2.13
dill==0.3.6
distro==1.8.0
frozenlist==1.3.3
greenlet==2.0.2
isort==5.12.0
lazy-object-proxy==1.9.0
loguru==0.6.0
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
netaddr==0.8.0
parse==1.18.0
platformdirs==2.6.2
pyaes==1.6.1
pylint==2.16.1
pylint-runner==0.6.0
Pyrogram==2.0.97
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
qiwipy==2.1.6
requests==2.25.1
six==1.15.0
SQLAlchemy==2.0.1
TgCrypto==1.2.5
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==3.7.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.2
win32-setctime==1.1.0
wrapt==1.14.1
yarl==1.8.2
yookassa==2.3.4
